+++ lib/src/controllers/th2_file_edit_controller.dart
@@ -913,6 +913,7 @@ abstract class TH2FileEditControllerBase with Store {
   @action
   void setShowImages(bool shouldShow) {
     _showImages = shouldShow;
+    userInteractionController.markStationPointNameCoordinateCacheDirty();
   }
 
   @action
@@ -1140,6 +1141,7 @@ abstract class TH2FileEditControllerBase with Store {
   void triggerAllElementsRedraw() {
     selectionController
         .clearSelectedElementsBoundingBoxAndSelectionHandleCenters();
+    userInteractionController.markStationPointNameCoordinateCacheDirty();
     _redrawTriggerAllElements++;
   }
 
@@ -1147,6 +1149,7 @@ abstract class TH2FileEditControllerBase with Store {
   void triggerSelectedElementsRedraw({bool setState = false}) {
     selectionController
         .clearSelectedElementsBoundingBoxAndSelectionHandleCenters();
+    userInteractionController.markStationPointNameCoordinateCacheDirty();
     _redrawTriggerSelectedElements++;
 
     if (setState) {
@@ -1156,6 +1159,7 @@ abstract class TH2FileEditControllerBase with Store {
 
   @action
   void triggerNonSelectedElementsRedraw() {
+    userInteractionController.markStationPointNameCoordinateCacheDirty();
     _redrawTriggerNonSelectedElements++;
   }
 
@@ -1183,6 +1187,7 @@ abstract class TH2FileEditControllerBase with Store {
 
   @action
   void triggerImagesRedraw() {
+    userInteractionController.markStationPointNameCoordinateCacheDirty();
     _redrawTriggerImages++;
   }